renaming server/version to server/storage. version was a bad name I picked late at night
